Added functionality in CategoriesController to not delete that category which contains posts.

+++ app/Http/Controllers/CategoriesController.php
@@ -102,6 +102,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted as it contains many posts.');
+            return redirect(route('categories.index'));
+        }
         $category->delete();
         session()->flash('success', 'Category deleted sucessfully!');
         return redirect(route('categories.index'));
